Code optimizations reported by static analysis

+++ src/faction.cpp
@@ -26,7 +26,7 @@ std::string invent_adj();
 
 constexpr inline unsigned long mfb( const int v )
 {
-    return 1 << v;
+    return static_cast<unsigned long>( 1 << v );
 }
 
 faction_template::faction_template()
